Extracted network functionality from thumbnail widget in to image widget
