+++ lib/features/post/widgets/widgets.dart
@@ -7,3 +7,4 @@ export 'content_widget/content_widget.dart';
 export 'video_player_widget/video_player_widget.dart';
 export 'video_player_controls_widget/video_player_controls_widget.dart';
 export 'date_widget/date_widget.dart';
+export 'image_widget/image_widget.dart';
